newsai-archive/ai-service/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
transformers
torch
pillow
openai-whisper
python-multipart
